+++ app/Filament/Resources/TokenResource/Pages/ListTokens.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\TokenResource\Pages;
+
+use App\Filament\Resources\TokenResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListTokens extends ListRecords
+{
+    protected static string $resource = TokenResource::class;
+}